[composeapp] renamed package rolls to rolledit
